Use systematic, non-repetitive way to get translations of select options
ApplicationRecord: Add i18n_keys attribute recording the key used to
access I18n translation values for each attribute of a model.

VueFormBuilder: Add smart_select method. This method creates a select in
a form and query automatically the necessary options.

Thanks to this we can delete a lot of repetitive helpers and reduce code
complexity.