BiochemicalAnalyses: Add input support
Input support is added in a very different way than the other analyses.
This is because BiochemicalAnalyses structure is triply nested.
AnalysisBiochemistry -> BiochemicalAnalyses -> BiochemicalAnalysisResults
Using the same middleware code would have been very tricky.

As a result, the code written in this commit is quite ad hoc and may be
a pain to maintain.

Issue: Preview is not updated. Will need to fix that later.