DualExaminationBaseDecorator: Refactor
We need to clearly differentiate dual examinations with normal dual
records.

Dual records (like the ears or the eyes) do not represent a full
examination. Dual examinations are a special case of dual records, as it
also includes the data relevant to an examination.

Here is the structure after the refactoring:

                                       ExternalEarDecorator          -> DualRecordDecorator  -> ApplicationDecorator
                                       ExternalEyeDecorator          -/
InternalLungExaminationDecorator    -> DualExaminationBaseDecorator  -/
InternalKidneyExaminationDecorator  -> DualExaminationBasesDecorator -/
                                       ExternalEarsDecorator         -> DualRecordsDecorator -> ApplicationCollectionDecorator
                                       ExternalEyesDecorator         -/
InternalLungsExaminationDecorator   -> DualExaminationBaseDecorator  -/
InternalKidneysExaminationDecorator -> DualExaminationBasesDecorator -/